==> notebooks/1. Create_MMDS_Courses.ipynb <==
#%pip install tabula-py

import pandas as pd
import numpy as np
import re
from tabula import read_pdf

## 1. Upload Data Source

#### Origin: Data Science Course Catalog

data_science_file = read_pdf("../data/origin/MK_MMDS_2022_23_neue_PO_29082022.pdf", pages="all")
data_science_file_copy = data_science_file.copy()
len(data_science_file)

# Keep tables with information (>2)
df = []
for table in data_science_file_copy:
    if len(table)>2:
        df.append(table)
len(df)

## 2. Data Preprocessing

Data Frame Cleaning
1. Remove ['Language', 'ECTS', 'Page'] from df[x]
2. Promote first row as header
3. Insert module category as a new feature

### Fundamentals Courses

df[1]
data_fundamentals = df[1].loc[:,~df[1].columns.isin(['Language', 'ECTS', 'Page','Unnamed: 2'])].copy()
data_fundamentals.columns = data_fundamentals.iloc[0]
data_fundamentals = data_fundamentals[2:]
data_fundamentals.columns = data_fundamentals.columns.fillna('module_code')
data_fundamentals['moduleCategory'] = 'Fundamentals'
data_fundamentals

### Data Management Courses

df[8]
data_management = df[8].loc[:,~df[8].columns.isin(['Language', 'ECTS', 'Page'])].copy()
data_management.columns = data_management.iloc[0]
data_management = data_management[2:]
data_management.columns = data_management.columns.fillna('module_code')
data_management['moduleCategory'] = 'Data Management'
data_management

### Data Analytics Courses

#Data Analystics
df[10]
data_analytics = df[10].loc[:,~df[10].columns.isin(['Language', 'ECTS', 'Page'])].copy()
data_analytics.columns = data_analytics.iloc[0]
data_analytics = data_analytics[2:]
data_analytics.columns = data_analytics.columns.fillna('module_code')
data_analytics['moduleCategory'] = 'Data Analytics'
data_analytics

### Responsible Data Science Courses

df[15]
data_responsible = df[15].loc[:,~df[15].columns.isin(['Language', 'ECTS', 'Page'])].copy()
data_responsible.columns = data_responsible.iloc[0]
data_responsible = data_responsible[2:]
data_responsible.columns = data_responsible.columns.fillna('module_code')
data_responsible['moduleCategory'] = 'Responsible Data Science'
data_responsible

### Consolidate Courses

final_frame = pd.concat([data_fundamentals,data_management,data_analytics,data_responsible],ignore_index=True,axis=0)

# Removing '\r' from the data
mask_r = final_frame['Name of Module'].str.contains('\r')
final_frame[mask_r]

# Final df cleaning
features = ['Name of Module','ECTS']
for col in features:
    final_frame.loc[mask_r,col] = final_frame.loc[mask_r,col].str.replace('\r',' ')
final_frame[mask_r]
for col in features:
    final_frame.loc[mask_r,col] = final_frame.loc[mask_r,col].str.replace('- ','')
final_frame[mask_r]

#Clean module_code
final_frame.module_code = final_frame.module_code.str.rstrip('**')
final_frame.module_code = final_frame.module_code.str.rstrip()
final_frame.module_code = final_frame.module_code.str.lstrip()
final_frame

final_frame

## 3. Save MMDS_Courses

# Save file
final_frame.to_csv('../data/processed/mmds_courses.csv',index=False)
