==> notebooks/2. Create_MMDS_Courses_Data.ipynb <==
import pandas as pd
import numpy as np
from tabula import read_pdf
import re
pd.set_option('display.max_rows', 200)
pd.set_option('display.max_colwidth', None)

## 1. Upload Data Sources

#### Origin: Bussiness Informatics Course Catalog

bi_file = read_pdf('../data/origin/MK_MSc_Wifo_2022_2023_29082022.pdf', pages='all')
bi_file_copy = bi_file.copy()

# Keep tables with information (>2)
wifo_df = []
for table in bi_file_copy:
    if len(table)>2:
        wifo_df.append(table)
len(wifo_df)

#### Origin: Data Science Course Catalog

data_science_file = read_pdf('../data/origin/MK_MMDS_2022_23_neue_PO_29082022.pdf', pages='all')
data_science_file_copy = data_science_file.copy()

# Keep tables with information (>2)
df_mmds = []
for table in data_science_file:
    if len(table)>2:
        df_mmds.append(table)
len(df_mmds)

#### Origin: Data Science List of Courses

courses = pd.read_csv('../data/processed/mmds_courses.csv')
courses

## 2. Data Preprocessing

### 2.1 Courses Identification

####  MMDS Courses in Bussiness Informatics Catalog

MMDS courses found in BI catalog

# Find the tables from BI where the information about the MMDS courses is located

# Create and index of the found courses and the table where the information is located
mask_na = courses.module_code.notna() # Ignore courses without a code (4 courses)
courses_ = courses[mask_na].copy()
wifo_df_idx = []
for item in courses_.module_code:
    for idx in range(len(wifo_df)):
        if item==wifo_df[idx].columns[0]:
            wifo_df_idx.append([idx,item])
            break
print('Index of MMDS courses found in BI catalog: (%s)\n'%len (wifo_df_idx), wifo_df_idx)
            
# List of the MMDS courses found in BI catalog
def extract_item(lst):
    return [item[1] for item in lst]
matched_courses_wifo = extract_item(wifo_df_idx)

MMDS courses not found in BI catalog

# List of the MMDS courses NOT found in BI catalog
unmatched_course_in_wifo = []
for item in courses_.module_code:
     if item not in matched_courses_wifo:
            unmatched_course_in_wifo.append(item)
            
print('List of MMDS courses NOT found in BI catalog: (%s)\n'%len (unmatched_course_in_wifo), unmatched_course_in_wifo)

####  MMDS Courses in MMDS Catalog

MMDS courses found in MMDS catalog

df_mmds_idx = []
for item in unmatched_course_in_wifo:
    for idx in range(len(df_mmds)):
        if item==df_mmds[idx].columns[0]:
            df_mmds_idx.append([idx,item])
            break
print('Index of MMDS courses found in MMDS catalog: (%s)\n'%len (df_mmds_idx), df_mmds_idx)

# List of the MMDS courses found in MMDS catalog
matched_courses_mmds = extract_item(df_mmds_idx)

MMDS courses not found in BI or MMDS catalog

# List of the MMDS courses NOT found in MMDS catalog
unmatched_course_in_wifo_mmds = []
for item in courses_.module_code:
     if item not in matched_courses_mmds+matched_courses_wifo:
            unmatched_course_in_wifo_mmds.append(item)
            
print('List of MMDS courses NOT found in BI or MMDS catalog: (%s)\n'%len (unmatched_course_in_wifo_mmds), unmatched_course_in_wifo_mmds)

### 2.2 Extraction of Courses Metadata from BI Catalog

#### Tables that are the start of courses metadata
Create list with tables that are the start of courses tables

# List of table indexes that are the start of a metadata tables of courses 
course_idx = []
for idx in range(len(wifo_df)):
    if bool(re.search('^[a-zA-Z]{1,3} ',wifo_df[idx].columns[0])):
        course_idx.append(idx)
print('List of table indexes of BI that are courses: (%s)\n'%len(course_idx), course_idx)

Fix format of tables with more than 2 columns

# List of tables with more than 2 columns ('Unnamed')
three_column_courses = []
for item in course_idx:
    if len(wifo_df[item].columns)>2:
        three_column_courses.append(item)
print('List of table indexes with more than 2 columns: \n', three_column_courses)


# Dropping extra columns from the tables that are in three_column_courses
drop_columns = ['Unnamed: 0','Unnamed: 1','Unnamed: 2']
for idx in three_column_courses:
    wifo_df[idx] = wifo_df[idx].loc[:,~wifo_df[idx].columns.isin(drop_columns)].copy()

# wifo_df[5].head() --- to validate if it work

Delete tables with just 1 column

# List of tables with just 1 column
one_column_courses = []
for item in course_idx:
    if len(wifo_df[item].columns)==1:
        one_column_courses.append(item)
print('List of table indexes with just 1 column: \n', one_column_courses)

# Remove the 22 idx from course index (not a course from mmds)
course_idx.remove(22)

Final list with tables that are the start of courses tables

print('List of table indexes of BI that are courses: (%s)\n'%len(course_idx), course_idx)

#### Tables that are not the start of courses metadata
Create list with tables that are not the start of courses tables

# List of table indexes that are not the start of a metadata tables of courses 
non_course_idx = []
for idx in range(len(wifo_df)):
    if not bool(re.search('^[a-zA-Z]{1,3} ',wifo_df[idx].columns[0])):
        non_course_idx.append(idx)
print('List of table indexes of BI that are not courses: (%s)\n'%len(non_course_idx), non_course_idx)

Fix format of tables with more than 2 columns, and delete tables without courses' detail

# List of tables with more than 2 columns ('Unnamed')
three_column_non_courses = []
for item in non_course_idx:
    if len(wifo_df[item].columns)>2:
        three_column_non_courses.append(item)
print('List of table indexes with more than 2 columns: \n', three_column_non_courses)


# Handling the data frames with more thant 3 columns
remove_non_course_idx = [0,1,14,15,16,17,65,66,67] # remove 0,1,14,15,16,17,65,66,67 indexes
for i in remove_non_course_idx:
    three_column_non_courses.remove(i)
    non_course_idx.remove(i)
    
for idx in three_column_non_courses:
    wifo_df[idx] = wifo_df[idx].loc[:,wifo_df[idx].columns[0:2]].copy()

Check tables with just 1 column

# List of tables with just 1 column
one_column_non_courses = []
for item in non_course_idx:
    if len(wifo_df[item].columns)==1:
        one_column_non_courses.append(item)
print('List of table indexes with just 1 column:\n', one_column_non_courses)


Final list with tables that are not the start of courses tables

print('List of table indexes of BI that are not courses: (%s)\n'%len(non_course_idx), non_course_idx)

### 2.3 Build the Metadata Dataset from MMDS Courses from BI Catalog

Join the frames that contain metadata about the MMDS courses

#Desired output with data frame
final_columns = ['module_code','module_name','property','property_value']
final_wifo_frame = pd.DataFrame(columns=final_columns)

#1st loop to append the consider course frames
for idx in range(len(course_idx)):
    course_frame = wifo_df[course_idx[idx]].copy()
    course_name = wifo_df[course_idx[idx]].columns[1]
    course_code = wifo_df[course_idx[idx]].columns[0]
    course_frame.columns = final_columns[2:]
    print(course_idx[idx])
    #if not last course
    if (idx+1)<len(course_idx):
        #2nd for loop to append the non course frames
        for nc_idx in non_course_idx:
            if (nc_idx>course_idx[idx]) & (nc_idx<course_idx[idx+1]):
                intermediate_frame = wifo_df[nc_idx].copy()
                intermediate_frame.columns = final_columns[2:]
                course_frame = pd.concat([course_frame,intermediate_frame],ignore_index = True)
                print('--',nc_idx)
    
    #for last course    
    else:
        for nc_idx in non_course_idx:
            if (nc_idx>course_idx[idx]):
                intermediate_frame = wifo_df[nc_idx].copy()
                intermediate_frame.columns = final_columns[2:]
                course_frame = pd.concat([course_frame,intermediate_frame],ignore_index = True)
                print('--',nc_idx)
    course_frame['module_code'] = course_code
    course_frame['module_name'] = course_name
    final_wifo_frame = pd.concat([final_wifo_frame,course_frame])
    print('appended')

display(final_wifo_frame.head())
print ('Size final_wifo_frame: ',len(final_wifo_frame))

### 2.4 Extraction of Courses Metadata from MMDS Catalog

#### Tables that are the start of courses metadata
Create list with tables that are the start of courses tables

# List of table indexes that are the start of a metadata tables of courses 
course_idx = []
for idx in range(len(df_mmds)):
    if bool(re.search('^[a-zA-Z]{1,3} ',df_mmds[idx].columns[0])):
        course_idx.append(idx)
print('List of table indexes of MMDS that are courses: (%s)\n'%len(course_idx), course_idx)

Fix format of tables with more than 2 columns

# List of tables with more than 2 columns ('Unnamed')
three_column_courses = []
for item in course_idx:
    if len(df_mmds[item].columns)>2:
        three_column_courses.append(item)
print('List of table indexes with more than 2 columns: \n', three_column_courses)


# Dropping extra columns from the tables that are in three_column_courses
drop_columns = ['Unnamed: 0','Unnamed: 1','Unnamed: 2']
for idx in three_column_courses:
    df_mmds[idx] = df_mmds[idx].loc[:,~df_mmds[idx].columns.isin(drop_columns)].copy()


Check tables with just 1 column

# List of tables with just 1 column
one_column_courses = []
for item in course_idx:
    if len(df_mmds[item].columns)==1:
        one_column_courses.append(item)
print('List of table indexes with just 1 column: \n', one_column_courses)

Final list with tables that are the start of courses tables

print('List of table indexes of MMDS that are courses: (%s)\n'%len(course_idx), course_idx)

#### Tables that are not the start of courses metadata
Create list with tables that are not the start of courses tables

# List of table indexes that are not the start of a metadata tables of courses 
non_course_idx = []
for idx in range(len(df_mmds)):
    if not bool(re.search('^[a-zA-Z]{1,3} ',df_mmds[idx].columns[0])):
        non_course_idx.append(idx)
print('List of table indexes of MMDS that are not courses: (%s)\n'%len(non_course_idx), non_course_idx)

Fix format of tables with more than 2 columns, and delete tables without courses' detail

# List of tables with more than 2 columns ('Unnamed')
three_column_non_courses = []
for item in non_course_idx:
    if len(df_mmds[item].columns)>2:
        three_column_non_courses.append(item)
print('List of table indexes with more than 2 columns: \n', three_column_non_courses)


# Handling the data frames with more thant 3 columns
remove_non_course_idx = [0,1,8,10,15,20] # remove 0,1,8,10,15,20 indexes
for i in remove_non_course_idx:
    three_column_non_courses.remove(i)
    non_course_idx.remove(i)
    
for idx in three_column_non_courses:
    df_mmds[idx] = df_mmds[idx].loc[:,df_mmds[idx].columns[0:2]].copy()

Check tables with just 1 column

# List of tables with just 1 column
one_column_non_courses = []
for item in non_course_idx:
    if len(df_mmds[item].columns)==1:
        one_column_non_courses.append(item)
print('List of table indexes with just 1 column:\n', one_column_non_courses)


Final list with tables that are not the start of courses tables

print('List of table indexes of MMDS that are not courses: (%s)\n'%len(non_course_idx), non_course_idx)

### 2.5 Build the Metadata Dataset from MMDS Courses from MMDS Catalog

Join the frames that contain metadata about the MMDS courses

#Desired output with data frame
final_columns = ['module_code','module_name','property','property_value']
final_mmds_frame = pd.DataFrame(columns=final_columns)

#1st loop to append the consider course frames
for idx in range(len(course_idx)):
    course_frame = df_mmds[course_idx[idx]].copy()
    course_name = df_mmds[course_idx[idx]].columns[1]
    course_code = df_mmds[course_idx[idx]].columns[0]
    course_frame.columns = final_columns[2:]
    print(course_idx[idx])
    #if not last course
    if (idx+1)<len(course_idx):
        #2nd for loop to append the non course frames
        for nc_idx in non_course_idx:
            if (nc_idx>course_idx[idx]) & (nc_idx<course_idx[idx+1]):
                intermediate_frame = df_mmds[nc_idx].copy()
                intermediate_frame.columns = final_columns[2:]
                course_frame = pd.concat([course_frame,intermediate_frame],ignore_index = True)
                print('--',nc_idx)
    
    #for last course    
    else:
        for nc_idx in non_course_idx:
            if (nc_idx>course_idx[idx]):
                intermediate_frame = df_mmds[nc_idx].copy()
                intermediate_frame.columns = final_columns[2:]
                course_frame = pd.concat([course_frame,intermediate_frame],ignore_index = True)
                print('--',nc_idx)
    course_frame['module_code'] = course_code
    course_frame['module_name'] = course_name
    final_mmds_frame = pd.concat([final_mmds_frame,course_frame])
    print('appended')

display(final_mmds_frame.head())
print ('Size final_wifo_frame: ',len(final_mmds_frame))

### 2.5 Combine MMDS Courses from MMDS Catalog and BI Catalog

Construct a df with all the triples necessary for the KG proposed.

final_frame = pd.concat([final_wifo_frame,final_mmds_frame])
print ('Size final_frame: ',len(final_frame))

# Keep just courses with code
mask_mmds_courses = final_frame.module_code.isin(list(courses_.module_code))
mmds_course_data = final_frame[mask_mmds_courses].copy()

# Fill nan of property
mmds_course_data['property'].ffill(axis=0, inplace=True)

# Dropping duplicate records
mmds_course_data.drop_duplicates(inplace=True)
print ('Size mmds_course_data: ',len(mmds_course_data))

# Grouping the property value
mmds_course_data.property_value = mmds_course_data.property_value.astype(str)
mmds_course_data_grouped = mmds_course_data.groupby(['module_code','module_name','property'])['property_value'].apply(lambda x: ','.join(x)).reset_index().copy()

#removing '\r' from the data
mask_r = mmds_course_data_grouped['property'].str.contains('\r')
features = list(mmds_course_data_grouped.columns)
for col in features:
    mask_r = mmds_course_data_grouped[col].str.contains('\r')
    mmds_course_data_grouped.loc[mask_r,col] = mmds_course_data_grouped.loc[mask_r,col].str.replace('\r',' ')
mmds_course_data_grouped[mask_r]
print ('Size mmds_course_data_grouped: ',len(mmds_course_data_grouped))


## 3. Save MMDS_Courses_Data

# Save file
mmds_course_data_grouped.to_csv('../data/processed/mmds_courses_data.csv',index=False)
