==> notebooks/3. Build_Couses_KG.ipynb <==
import pandas as pd
import numpy as np
import re
pd.set_option('display.max_rows', 200)
pd.set_option('display.max_colwidth', None)

import rdflib
from urllib.parse import quote
from rdflib import Graph, URIRef, Literal, BNode, Namespace
from rdflib.namespace import FOAF, RDF, OWL, RDFS, XSD

## 1. Upload Data Sources

#### MMDS_Courses

df_mmds_courses = pd.read_csv('../data/processed/mmds_courses.csv')
df_mmds_courses.ECTS = np.where(df_mmds_courses.ECTS.str.isdigit(), df_mmds_courses.ECTS, '18')
df_mmds_courses.ECTS = df_mmds_courses.ECTS.astype(int)

display(df_mmds_courses.head())
print ('Size df_mmds_courses: ',len(df_mmds_courses))

df_mmds_courses[df_mmds_courses.module_code.isna()]

mmds_modules = df_mmds_courses.moduleCategory.value_counts().rename_axis('module').reset_index(name="counts")
mmds_modules

#### MMDS_Courses_Data

mmds_courses_data = pd.read_csv('../data/processed/mmds_courses_data.csv')
display(mmds_courses_data.head())
print ('Size mmds_courses_data: ',len(mmds_courses_data))

#Nulls in property_value column
len(mmds_courses_data[mmds_courses_data.property_value.isna()])

## 2. URIs Creation

#### Programs : https://www.wim.uni-mannheim.de/en/academics/programs/

        - Base URL : https://www.uni-mannheim.de/en/academics/
        - MMDS : https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/
        - BI.  : https://www.uni-mannheim.de/en/academics/programs/masters-program-in-business-informatics/
        - Courses : https://www.uni-mannheim.de/en/academics/courses/

baseURL = Namespace('https://www.uni-mannheim.de/en/academics/')
mmdsModuleURL = Namespace(baseURL+'programs/mannheim-master-in-data-science/modules/')
mmdsURL = Namespace(baseURL+'programs/mannheim-master-in-data-science/')
courseURL = Namespace(baseURL+'courses/')
propertyURL = Namespace(baseURL+'property/')
professorURL = Namespace(baseURL+'professor/')

def create_uri(name:str,baseURL):
    '''Takes a string and returns a valid namespace URI'''
    quoted = quote(name.replace(" ", "_"))
    return baseURL[quoted]

### 2.1 Module URIs

mmds_modules['moduleURI'] = mmds_modules.module.apply(lambda x: create_uri(x,mmdsModuleURL))

display (mmds_modules)
print('Example: ',mmds_modules.moduleURI[0])

### 2.2 Course URIs

# Eliminate instances without code
mmds_courses = df_mmds_courses[df_mmds_courses.module_code.notna()].copy()
# Create URIs
mmds_courses['courseURI'] = mmds_courses.module_code.str.lower()+' '+mmds_courses['Name of Module'].str.lower()
mmds_courses['courseURI'] = mmds_courses.courseURI.apply(lambda x: create_uri(x,courseURL))

display (mmds_courses.head())
print('Example: ',mmds_courses['courseURI'][0])

### 2.3 Property URIs

#### Properties selection

mmds_course_property = mmds_courses_data['property'].value_counts().rename_axis('property').reset_index(name="counts").copy()
mmds_course_property = mmds_course_property[mmds_course_property.counts>16]
mmds_course_property.head()

Selected properties

property_list = ['Duration of assessment','Prerequisites','Form of assessment','Aim of module',
                 'Workload','Lecturer','Duration of module','Semester','Methods',
                 'Range of application','Admission requirements']
print (property_list)

Filtrate dataframe, including only these properties

mmds_courses_data_ = mmds_courses_data[mmds_courses_data.property.isin(property_list)].copy()
print ('Size mmds_courses_data_: ',len(mmds_courses_data_))

Create course URI, property URI, and professor URI

#create course URI
mmds_courses_data_
mmds_courses_data_['courseURI'] = mmds_courses_data_.module_code.str.lower()+' '+mmds_courses_data_.module_name.str.lower()
mmds_courses_data_['courseURI'] = mmds_courses_data_.courseURI.apply(lambda x: create_uri(x,courseURL))

#create property URI
mmds_courses_data_['propertyURI'] = mmds_courses_data_.property.str.lower()
mmds_courses_data_['propertyURI'] = mmds_courses_data_.propertyURI.apply(lambda x: create_uri(x,propertyURL))
mmds_courses_data_.head()

#Unique list of properties
propertyURI = list(mmds_courses_data_.propertyURI.unique())
propertyURI

### 2.4 Lecturer URIs

#create professor URI
lecturere_mask = mmds_courses_data_.property=='Lecturer'
mmds_courses_data_.loc[lecturere_mask,'property_value'] = mmds_courses_data_.loc[lecturere_mask,'property_value'].str.replace('.','',regex=True).str.replace('; ',',',regex=True).str.replace(',nan','',regex=True)
mmds_courses_data_.loc[lecturere_mask,'property_value'] = mmds_courses_data_.loc[lecturere_mask,'property_value'].str.replace('Lecturer at the host university','LHU',regex=True).str.replace('Prof ','',regex=True).str.replace('Dr ','',regex=True).str.replace('Junior ','',regex=True).str.replace('Professor ','',regex=True).str.replace('Simone Paolo Ponzetto','Simone Ponzetto',regex=True)
mmds_courses_data_.loc[lecturere_mask,'property_value'] = mmds_courses_data_.loc[lecturere_mask,'property_value'].str.split(",")
mmds_courses_data_ = mmds_courses_data_.explode('property_value')
mmds_courses_data_.loc[lecturere_mask,'property_value'] = mmds_courses_data_[lecturere_mask].property_value.apply(lambda x: create_uri(x,professorURL))

print ('Size mmds_courses_data_: ',len(mmds_courses_data_))

#Unique list of Lecturers
lecturerURI = list(mmds_courses_data_.loc[lecturere_mask,'property_value'].unique())
lecturerURI

## 3. KG Building

baseURL = Namespace('https://www.uni-mannheim.de/en/academics/')
mmdsModuleURL = Namespace(baseURL+'programs/mannheim-master-in-data-science/modules/')
mmdsURL = Namespace(baseURL+'programs/mannheim-master-in-data-science/')
courseURL = Namespace(baseURL+'courses/')
propertyURL = Namespace(baseURL+'property/')
professorURL = Namespace(baseURL+'professor/')

g = Graph()
g.bind("foaf", FOAF)
g.bind("courseKG", baseURL)
g.bind("prop",propertyURL)
g.bind("owl",OWL)
g.bind("xsd",XSD)

baseURL.programs

### 3.1 Ontology

#### Classes

g.add((baseURL.programs, RDF.type, RDFS.Class))
g.add((baseURL.module, RDF.type, RDFS.Class))
g.add((baseURL.course, RDF.type, RDFS.Class))
g.add((baseURL.people, RDF.type, RDFS.Class ))
g.add((baseURL.people, RDFS.subClassOf, FOAF.Person))
g.add((baseURL.professor,RDFS.subClassOf,baseURL.people))

#### Properties

# Set all properties as DatatypeProperty
for prop in propertyURI:
    g.add((prop, RDF.type, OWL.DatatypeProperty))
    g.add((prop, RDFS.domain, baseURL.course))
    g.add((prop, RDFS.range, XSD.string))
    
# Change property lecturer to ObjectProperty
g.set((propertyURL.lecturer, RDF.type, OWL.ObjectProperty))
g.set((propertyURL.lecturer, RDFS.domain, baseURL.course))
g.set((propertyURL.lecturer, RDFS.range, baseURL.professor))

# isCourseOf property
g.add((propertyURL.isCourseOf, RDF.type, OWL.ObjectProperty))
g.add((propertyURL.isCourseOf, RDFS.domain, baseURL.course))
g.add((propertyURL.isCourseOf, RDFS.range, baseURL.module))

# isModuleOf property
g.add((propertyURL.isModuleOf, RDF.type, OWL.ObjectProperty))
g.add((propertyURL.isModuleOf, RDFS.domain, baseURL.module))
g.add((propertyURL.isModuleOf, RDFS.range, baseURL.programs))

# courseCode property
g.add((propertyURL.courseCode, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.courseCode, RDFS.domain, baseURL.course))
g.add((propertyURL.courseCode, RDFS.range, XSD.string))

# courseName property
g.add((propertyURL.courseName, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.courseName, RDFS.domain, baseURL.course))
g.add((propertyURL.courseName, RDFS.range, XSD.string))

# ects property
g.add((propertyURL.ects, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.ects, RDFS.domain, baseURL.course))
g.add((propertyURL.ects, RDFS.range, XSD.integer))

# language property
g.add((propertyURL.language, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.language, RDFS.domain, baseURL.course))
g.add((propertyURL.language, RDFS.range, XSD.string))

# offering property
g.add((propertyURL.offering, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.offering, RDFS.domain, baseURL.course))
g.add((propertyURL.offering, RDFS.range, XSD.string))

# programName property
g.add((propertyURL.programName, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.programName, RDFS.domain, baseURL.programs))
g.add((propertyURL.programName, RDFS.range, XSD.string))

# moduleName property
g.add((propertyURL.moduleName, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.moduleName, RDFS.domain, baseURL.module))
g.add((propertyURL.moduleName, RDFS.range, XSD.string))

# professorName property
g.add((propertyURL.professorName, RDF.type, OWL.DatatypeProperty))
g.add((propertyURL.professorName, RDFS.domain, baseURL.professor))
g.add((propertyURL.professorName, RDFS.range, XSD.string))


### 3.2 Assertion - Box

#### Programs

g.add((URIRef(mmdsURL), RDF.type, baseURL.programs))
g.add((URIRef(mmdsURL), propertyURL.programName, Literal('mannheim-master-in-data-science')))

#### Modules

Define modules and connect them with literal values (moduleName)

for row in mmds_modules.index:
    g.add((URIRef(mmds_modules.loc[row,'moduleURI']), RDF.type, baseURL.module))
    g.add((mmds_modules.loc[row,'moduleURI'], propertyURL.moduleName, Literal(mmds_modules.loc[row,'module'])))

Connect modules with programs

for item in mmds_modules.moduleURI.iteritems():
    g.add((item[1],propertyURL.isModuleOf,URIRef(mmdsURL)))

#### Professors

for profURI in lecturerURI:
    g.add((profURI, RDF.type, baseURL.professor))
    g.add((profURI, propertyURL.professorName, Literal(str(profURI).replace(str(professorURL),''))))

Connect to public KG

# OWL sameAs links to other (public) knowledge graphs
sameAs = {
    'https://www.uni-mannheim.de/en/academics/professor/Ursula_Rost' : ['https://dblp.org/pid/r/URost'],
    'https://www.uni-mannheim.de/en/academics/professor/Heiko_Paulheim': ['https://dblp.org/pid/39/4064', 'http://www.wikidata.org/entity/Q23709849'],
    'https://www.uni-mannheim.de/en/academics/professor/Marcus_Kessel' : ['https://dblp.org/pid/133/2140'],
    'https://www.uni-mannheim.de/en/academics/professor/Colin_Atkinson' : ['https://dblp.org/pid/a/ColinAtkinson', 'https://www.wikidata.org/entity/Q40100239'],
    'https://www.uni-mannheim.de/en/academics/professor/Guido_Moerkotte' : ['https://dblp.org/pid/m/GuidoMoerkotte','https://www.wikidata.org/entity/Q40472650'],
    'https://www.uni-mannheim.de/en/academics/professor/Matthias_Krause' : ['https://dblp.org/pid/90/2046-1','https://www.wikidata.org/entity/Q102110274'],
    'https://www.uni-mannheim.de/en/academics/professor/Rainer_Gemulla': ['https://dblp.org/pid/32/5357', 'http://www.wikidata.org/entity/Q39160786'],
    'https://www.uni-mannheim.de/en/academics/professor/Frederik_Armknecht' : ['https://dblp.org/pid/65/6856', 'https://www.wikidata.org/entity/Q102110273'],
    'https://www.uni-mannheim.de/en/academics/professor/Roland_Lei%C3%9Fa': ['https://dblp.org/pid/05/10957','https://www.wikidata.org/entity/Q108751937'],
    'https://www.uni-mannheim.de/en/academics/professor/Christian_Bizer': ['https://dblp.org/pid/b/ChristianBizer', 'http://www.wikidata.org/entity/Q17744291'],
    'https://www.uni-mannheim.de/en/academics/professor/Heiner_Stuckenschmidt': ['https://dblp.org/pid/50/5895', 'http://www.wikidata.org/entity/Q38721681'],
    'https://www.uni-mannheim.de/en/academics/professor/Markus_Strohmaier' : ['https://dblp.org/pid/01/6659', 'https://www.wikidata.org/entity/Q48612584'],
    'https://www.uni-mannheim.de/en/academics/professor/Simone_Ponzetto': ['https://dblp.org/pid/04/2532', 'http://www.wikidata.org/entity/Q28578432']
    }

for original, extensions in sameAs.items():
    for extension in extensions:
        g.add((URIRef(original), OWL.sameAs, URIRef(extension)))
        

#### Courses

for courseREF in mmds_courses.courseURI:
    g.add((URIRef(courseREF), RDF.type, baseURL.course))

Connect courses with literal values and object instances

def add_course_details(course_data):
    #reset the index of dataframe to avoid indexing errors
    course_data = course_data.reset_index()
    #counter for row number
    row=0
    num_rows = len(course_data.index)
    
    #for each row in the dataframe
    while row<num_rows:
        #add the triple
        sub, prop, val = (course_data.loc[row,'courseURI'],course_data.loc[row,'propertyURI'], course_data.loc[row,'property_value'])
        
        #check if value is null
        if not pd.isnull((val)):
            #check if the property is lecturer
            if course_data.loc[row,'property']=='Lecturer':
                #link the lecturer object to course
                g.add((sub, prop, val))
            else:
                #link the literal to the course
                g.add((sub, prop, Literal(val.lower())))
        row = row+1

add_course_details(mmds_courses_data_)

Connect courses with modules and more literal values (courseCode, courseName)

def link_courses_modules(mmds_courses):
    
    #reset the index of dataframe to avoid indexing errors
    mmds_courses = mmds_courses.reset_index()
    #counter for row number
    row=0
    num_rows = len(mmds_courses.index)
    
    #for each row in the dataframe
    while row<num_rows:
        #create a mapping dictionary for adding predicates
        map_dict = {"courseCode":mmds_courses.loc[row,"module_code"],
                    "courseName":mmds_courses.loc[row,"Name of Module"],
                    "ects":mmds_courses.loc[row,"ECTS"],
                    "language":mmds_courses.loc[row,"Language"],
                    "offering":mmds_courses.loc[row,"Offered"],
                    "modURI":create_uri(mmds_courses.loc[row,"moduleCategory"],mmdsModuleURL),
                    "courseURI":mmds_courses.loc[row,"courseURI"]}
        #add triples to link courses with modules
        g.add((map_dict["courseURI"], propertyURL.isCourseOf, map_dict["modURI"]))
        #add courseCode
        g.add((map_dict["courseURI"], propertyURL.courseCode, Literal(map_dict["courseCode"])))
        #add courseName
        g.add((map_dict["courseURI"], propertyURL.courseName, Literal(map_dict["courseName"])))
        #add ects
        g.add((map_dict["courseURI"], propertyURL.ects, Literal(int(map_dict["ects"]))))
        #add language
        g.add((map_dict["courseURI"], propertyURL.language, Literal(map_dict["language"])))
        #add offering 
        g.add((map_dict["courseURI"], propertyURL.offering, Literal(map_dict["offering"])))
            
        row=row+1

link_courses_modules(mmds_courses)

## 4. Save File I/O

s = g.serialize(format="ttl")
g.serialize(destination="../data/kg/mmds_courses_kg.ttl", format="ttl", encoding="utf-8") ;

#print(s)
