==> notebooks/4. Merge_Couses_&_Dws_KG.ipynb <==
import pandas as pd
import numpy as np
import re
pd.set_option('display.max_rows', 200)
pd.set_option('display.max_colwidth', None)

import rdflib
from urllib.parse import quote
from rdflib import Graph, URIRef, Literal, BNode, Namespace
from rdflib.namespace import FOAF, RDF, OWL, RDFS, XSD

## 1. Upload KGs

#### MMDS Courses KG

g_c = Graph()
g_c.parse('../data/kg/mmds_courses_kg.ttl')

#### Data Web Science KG

g_dws = Graph()
g_dws.parse('../data/kg/unima_dws_kg.ttl', format='turtle')

## 2. Combie KGs

g = g_c + g_dws

OWL sameAs links between the professors from MMDS Courses KG and Data Web Science KG

sameAs = {
    'https://www.uni-mannheim.de/en/academics/professor/Heiko_Paulheim': 'https://www.uni-mannheim.de/dws/people/professors/prof-dr-heiko-paulheim/',
    'https://www.uni-mannheim.de/en/academics/professor/Rainer_Gemulla': 'https://www.uni-mannheim.de/dws/people/professors/prof-dr-rainer-gemulla/',
    'https://www.uni-mannheim.de/en/academics/professor/Christian_Bizer': 'https://www.uni-mannheim.de/dws/people/professors/prof-dr-christian-bizer/',
    'https://www.uni-mannheim.de/en/academics/professor/Heiner_Stuckenschmidt': 'https://www.uni-mannheim.de/dws/people/professors/prof-dr-heiner-stuckenschmidt/',
    'https://www.uni-mannheim.de/en/academics/professor/Simone_Ponzetto': 'https://www.uni-mannheim.de/dws/people/professors/prof-dr-simone-paolo-ponzetto/'
    }

for mmds_professors, dws_professors in sameAs.items():
    g.add((URIRef(mmds_professors), OWL.sameAs, URIRef(dws_professors)))
    

s = g.serialize(format='ttl')
g.serialize(destination='../data/kg/mmds_courses_final_kg.ttl', format='ttl', encoding='utf-8')

#print(s)
