==> notebooks/5. SPARQL_Queries.ipynb <==
from rdflib import Graph, URIRef, Literal, BNode, Namespace
import re

%reload_ext jupyter-rdfify

## 1. Upload KG

g = Graph()
g.parse('../data/kg/mmds_courses_final_kg.ttl')

s = g.serialize(format='ttl')
#print(s)

## 2. SPARQL Queries

What are the modules of MMDS program?

modules_query = """
SELECT ?module_name
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> ;
            prop:moduleName ?module_name .
}"""

query = g.query(modules_query)
for row in query:
    print(f"{row.module_name}")

What are the courses and modules offered in MMDS?

modules_courses_query = """
SELECT ?module_n ?course_n
WHERE {
    ?module prop:isModuleOf <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> ;
            prop:moduleName ?module_n .
    ?course prop:isCourseOf  ?module ;
            prop:courseName ?course_n .
    
}"""

query = g.query(modules_courses_query)
for row in query:
    print(f"{row.module_n} - {row.course_n}")
    

What are the courses in Data Management for MMDS students?

dm_query = """
SELECT ?course_name
WHERE {
    ?course prop:isCourseOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/modules/Data_Management> ;
            prop:courseName ?course_name .
}"""

query = g.query(dm_query)
for row in query:
    print(f"{row.course_name}")


Which Data Analytics courses are taught in German?

da_courses_german_query = """
SELECT ?course_name ?lang
WHERE {
    ?course prop:isCourseOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/modules/Data_Analytics> ;
            prop:courseName ?course_name ;
            prop:language ?lang .
    FILTER ( regex(?lang,"G") )
    
}"""

query = g.query(da_courses_german_query)
for row in query:
    print(f"{row.course_name} - {row.lang}")
    

Which are the courses that have less than 6 ECTS?

courses_ects_query = """
SELECT ?course_name ?ects
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:ects ?ects .
    FILTER ( ?ects < 6 )
      
}"""

query = g.query(courses_ects_query)
for row in query:
    print(f"{row.course_name} - {row.ects}")

Which courses have an oral exam?

courses_form_assessment_query = """
SELECT ?course_name ?assessment
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:form_of_assessment ?assessment .
    FILTER ( regex(?assessment,"oral exam") )
      
}"""

query = g.query(courses_form_assessment_query)
for row in query:
    print(f"{row.course_name} - {row.assessment}")
    

Which courses have an examination of 60 minutes?

courses_dur_assessment_query = """
SELECT ?course_name ?assessment
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:duration_of_assessment ?assessment .
    FILTER ( regex(?assessment,"60") )
      
}"""

query = g.query(courses_dur_assessment_query)
for row in query:
    print(f"{row.course_name} - {row.assessment}")
    

Which courses are not recommended for a freshman?

courses_sm_query = """
SELECT ?course_name ?sm
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:semester ?sm .
    FILTER ( regex(?sm,"^((?!1).)*$")  )
      
}"""

query = g.query(courses_sm_query)
for row in query:
    sm_value = ', '.join([s for s in re.findall(r'\d+',row.sm)])
    print(f"{row.course_name} recommended for {sm_value} semester")
    

What courses are offered on FSS?

courses_off_query = """
SELECT ?course_name ?off
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:offering ?off .
    FILTER ( regex(?off,"FSS") )
      
}"""

query = g.query(courses_off_query)
for row in query:
    print(f"{row.course_name} - {row.off}")
    

Which courses are related to machine learning?

courses_ml_query = """
SELECT ?course_name
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:aim_of_module ?aim ;
            prop:prerequisites ?pre .
    FILTER ( regex(?aim,"machine learning") || regex(?pre,"machine learning") )
      
}"""

query = g.query(courses_ml_query)
for row in query:
    print(f"{row.course_name}")
    

Which courses use graphs or networks?

courses_graphs_query = """
SELECT ?course_name
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:aim_of_module ?aim ;
            prop:prerequisites ?pre .
    FILTER ( regex(?aim,"graphs") || regex(?pre,"graphs") || regex(?aim,"networks") || regex(?pre,"networks")  )
      
}"""

query = g.query(courses_graphs_query)
for row in query:
    print(f"{row.course_name}")
    

Which courses are about NLP?

courses_nlp_query = """
SELECT ?course_name
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:aim_of_module ?aim ;
            prop:prerequisites ?pre .
    FILTER ( regex(?aim,"nlp") || regex(?pre,"nlp") )
      
}"""

query = g.query(courses_nlp_query)
for row in query:
    print(f"{row.course_name}")
    

Which courses have as prerequisite a background in data mining?

courses_pre_query = """
SELECT ?course_name ?pre
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:courseName ?course_name ;
            prop:prerequisites ?pre .
    FILTER ( regex(?pre,"data mining") )
      
}"""

query = g.query(courses_pre_query)
for row in query:
    print(f"{row.course_name}")
    

What is the mail of the professors that teach courses related to DL?

course_prof_query = """
SELECT DISTINCT ?course_n ?prof_name ?prof_mail
WHERE {
    ?module prop:isModuleOf <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/modules/Data_Analytics> ;
            prop:courseName ?course_n ;
            prop:lecturer ?professor ;           
            prop:aim_of_module ?aim ;
            prop:prerequisites ?pre .      
    ?professor prop:professorName ?prof_name ;
                ns1:sameAs ?prof_dws .
    ?prof_dws a unima:professor .
    OPTIONAL {?prof_dws foaf:mbox ?prof_mail }
    FILTER ( regex(?aim,"deep learning") || regex(?pre,"deep learning") )
    
}"""

query = g.query(course_prof_query)
for row in query:
    print(f"{row.course_n} given by {row.prof_name}, email: {row.prof_mail}")
    

Which professor teaches the most courses?

lecturer_courses_query = """
SELECT DISTINCT ?name (count(?course) as ?count)
WHERE {
    ?module prop:isModuleOf  <https://www.uni-mannheim.de/en/academics/programs/mannheim-master-in-data-science/> .
    ?course prop:isCourseOf ?module ;
            prop:lecturer ?professor .
    ?professor prop:professorName ?name .
}
GROUP BY ?professor
ORDER BY DESC (?count)
"""

query = g.query(lecturer_courses_query)
for row in query:
    print(f"{row.name} teachs {row['count']} lectures")
    

Where did the professors graduate from?

%%time
wiki_uni_query = """
SELECT DISTINCT ?prof_name ?uni_label
WHERE {
    
    ?professor a courseKG:professor .
    ?professor prop:professorName ?prof_name .
    ?professor ns1:sameAs ?wikiperson .
    SERVICE <https://query.wikidata.org/bigdata/namespace/wdq/sparql> {
        ?wikiperson wdt:P69 ?uni .
        OPTIONAL {
            ?uni rdfs:label ?uni_label .
            filter (langMatches(lang(?uni_label),"EN"))
        }  
    } 
}
GROUP BY ?prof_name ?uni
"""
    

g.bind("wdt", Namespace("http://www.wikidata.org/prop/direct/"))
qwikidata = g.query(wiki_uni_query)
for row in qwikidata:
    print(f"{row.prof_name} \t {row.uni_label} ")

Which professor had more publications in the last year?
 - This question would be answered when DBPL provide a proper SPARQL endpoint in the near future. (https://blog.dblp.org/2022/03/02/dblp-in-rdf/)
